web apis for payments

+++ app/Http/Controllers/PlanRequestController.php
@@ -2,12 +2,15 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Company;
 use App\Models\Order;
 use App\Models\Plan;
 use App\Models\PlanRequest;
 use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
+use Illuminate\Support\Facades\Http;
 
 class PlanRequestController extends Controller
 {
@@ -19,9 +22,9 @@ class PlanRequestController extends Controller
     public function index()
     {
 
-        if (Auth::user()->type == 'super admin') {
+        if (Auth::user()->type == 'super admin' || Gate::check('Manage Plan Request')) {
             $plan_requests = PlanRequest::all();
-
+            // dd($plan_requests);
             return view('plan_request.index', compact('plan_requests'));
         } else {
             return redirect()->back()->with('error', __('Permission Denied.'));
@@ -52,67 +55,61 @@ class PlanRequestController extends Controller
      * @plan_id = Plan ID encoded
      * @duration = what duration is selected by user while request
     */
-    public function userRequest($plan_id)
-    {
-        $objUser = Auth::user();
+    //-----------------------------------------------------------------------------
+    //             this method is in api connecting global and tenent admins
+    //-----------------------------------------------------------------------------
+    // public function userRequest($plan_id)
+    // {
+    //     $objUser = Auth::user();
 
-        if ($objUser->requested_plan == 0) {
-            $planID = \Illuminate\Support\Facades\Crypt::decrypt($plan_id);
+    //     if ($objUser->requested_plan == 0) {
+    //         $planID = \Illuminate\Support\Facades\Crypt::decrypt($plan_id);
 
-            if (!empty($planID)) {
-                PlanRequest::create([
-                    'user_id' => $objUser->id,
-                    'plan_id' => $planID,
+    //         if (!empty($planID)) {
+    //             PlanRequest::create([
+    //                 'user_id' => $objUser->id,
+    //                 'plan_id' => $planID,
 
-                ]);
+    //             ]);
 
-                // Update User Table
+    //             // Update User Table
 
-                $objUser['requested_plan'] = $planID;
-                $objUser->update();
+    //             $objUser['requested_plan'] = $planID;
+    //             $objUser->update();
 
-                return redirect()->back()->with('success', __('Request Send Successfully.'));
-            } else {
-                return redirect()->back()->with('error', __('Something went wrong.'));
-            }
-        } else {
-            return redirect()->back()->with('error', __('You already send request to another plan.'));
-        }
-    }
-
+    //             return redirect()->back()->with('success', __('Request Send Successfully.'));
+    //         } else {
+    //             return redirect()->back()->with('error', __('Something went wrong.'));
+    //         }
+    //     } else {
+    //         return redirect()->back()->with('error', __('You already send request to another plan.'));
+    //     }
+    // }
+    //--------------------------------------------------------------------------------------------------
     /*
      * @id = Project ID
      * @response = 1(accept) or 0(reject)
     */
     public function acceptRequest($id, $response)
     {
-        if (Auth::user()->type == 'super admin') {
+        if (Auth::user()->type == 'super admin' || Gate::check('Manage Plan Request')) {
+
             $plan_request = PlanRequest::find($id);
-            if (!empty($plan_request)) {
-                $user = User::find($plan_request->user_id);
+            $company = Company::where('id', $plan_request->company_id)->first();
+
+            if (!empty($plan_request) && $company) {
+                // $user = User::find($plan_request->user_id);
 
                 if ($response == 1) {
-                    $user->requested_plan = $plan_request->plan_id;
-                    $user->plan           = $plan_request->plan_id;
-                    $user->requested_plan = '0';
-                    $user->save();
 
                     $plan       = Plan::find($plan_request->plan_id);
-                    $assignPlan = $user->assignPlan($plan_request->plan_id, $plan_request->duration);
                     $price      = $plan->price;
 
-                    if ($assignPlan['is_success'] == true && !empty($plan)) {
-                        if (!empty($user->payment_subscription_id) && $user->payment_subscription_id != '') {
-                            try {
-                                $user->cancel_subscription($user->id);
-                            } catch (\Exception $exception) {
-                                \Log::debug($exception->getMessage());
-                            }
-                        }
-
+                    if (!empty($plan)) {
                         $orderID = strtoupper(str_replace('.', '', uniqid('', true)));
-                        Order::create([
+                        $new_order = Order::create([
                             'order_id' => $orderID,
+                            'company_id' => $plan_request->company_id,
                             'name' => null,
                             'email' => null,
                             'card_number' => null,
@@ -123,27 +120,43 @@ class PlanRequestController extends Controller
                             'price' => $price,
                             'price_currency' => !empty(env('CURRENCY_CODE')) ? env('CURRENCY_CODE') : 'usd',
                             'txn_id' => '',
-                            'payment_type' => __('Manually Upgrade By Super Admin'),
+                            'payment_type' => __('Manually Upgrade By The Global Admin'),
                             'payment_status' => 'succeeded',
                             'receipt' => null,
-                            'user_id' => $user->id,
+                            'user_id' => $plan_request->user_id,
                         ]);
+                         
+                        if ($new_order) {
+                            
+                            $put_accept_data = Http::post($company->url . '/api/plan_request_accept_or_reject/' . $id . '/' . $response, [
+                                'new_order' => json_encode($new_order)
+                            ]);
 
-                        $plan_request->delete();
-
-                        return redirect()->back()->with('success', __('Plan successfully upgraded.'));
-                    } else {
+                            $put_accept_data = $put_accept_data->json();
+                            // dd($put_accept_data);
+                            if ($put_accept_data['status'] == 200) {
+                                $plan_request->delete();
+                                return redirect()->back()->with('success', __($put_accept_data['message']));
+                            } else {
+                                return redirect()->back()->with('error', __($put_accept_data['message']));
+                            }
+                        } 
                         return redirect()->back()->with('error', __('Plan fail to upgrade.'));
+                    } else {
+                        return redirect()->back()->with('error', __('Plan Not found.'));
                     }
                 } else {
-                    // $user->update(['requested_plan' => '0']);
-                    
-                    $user['requested_plan'] = 0;
-                    $user->update();
+                    $put_accept_data = Http::post($company->url . '/api/plan_request_accept_or_reject/' . $id . '/' . $response);
 
-                    $plan_request->delete();
+                    $put_accept_data = $put_accept_data->json();
 
-                    return redirect()->back()->with('success', __('Request Rejected Successfully.'));
+                    // dd($put_accept_data['status'] );
+                    if ($put_accept_data['status'] == 200) {
+                        $plan_request->delete();
+                        return redirect()->back()->with('success', __($put_accept_data['message']));
+                    } else {
+                        return redirect()->back()->with('error', __($put_accept_data['message']));
+                    }
                 }
             } else {
                 return redirect()->back()->with('error', __('Something went wrong.'));
@@ -168,7 +181,7 @@ class PlanRequestController extends Controller
 
         return redirect()->back()->with('success', __('Request Canceled Successfully.'));
     }
-    
+
     public function show(PlanRequest $planRequest)
     {
     }
